Redo buildEnv and add buildEnvCopy + buildFHSEnv
`buildEnv` populates its output with symlinks to the proper
`/nix/store/...` paths. However, there are cases in which /nix/store
will be inaccessible, such as inside the bwrap container pressure-vessel
creates for proton 5.13 Steam games. buildEnvCopy was added for this.

The way it works is by replacing the symlink operation in `builder.pl`
with a directory-recursive copy operation. Additionally, a `postBuild`
task is added, in which symlinks that lie outside `$out` are replaced
with the file they point to. To illustrate this, imagine a typical
dynamic library symlink chain:

```
libfoo.so -> libfoo.so.0 -> **libfoo.so.0.0.0** ->　/nix/store/.../lib/libfoo.so.0.0.0.
```

`libfoo.so` and `libfoo.so.0` point inside `$out`, so they will be left
as links. `libfoo.so.0.0.0` points to a library outside of `$out`, so it
will be replaced with said library.

Finally, ELF files are modified with patchelf to use the dynamic
interpreter inside `$out` and use their own directory or `../lib{64,32}`
as RUNPATH.

`buildFHSEnv` is added as a convenience from the `buildFHSUserEnv` to
just create the root, without hiding it behind a chroot/bwrap script.