steam: Allow to run soldier runtime games
The meat of this commit is the steamrt package. It creates a path with
everything pressure-vessel needs copied inside, graphics drivers and
libraries. It does not, however, put it in a bwrap, chroot, or container
of any kind; pressure-vessel will already do that. That's where the
other changes come in: for pressure-vessel to use this store path as its
sysroot, we need to set `PRESSURE_VESSEL_GRAPHICS_PROVIDER` before
calling it, or, more easily, before calling Steam.

Additionally, we need to know which drivers this system uses. By
default, only open-source mesa drivers are included. When toggling
`programs.steam.enable` to `true` in a config, the steamrt package
passed to steam will be overriden to receive the actual drivers as
inputs, and it's that steam, with steamrt produced from proper drivers,
that will be placed `environment.systemPackages`. As such, it's not
recommended to put Steam or steamrt there manually.