glibc: use ld cache from /run/current-system/sw/etc
Previously, a path inside the derivation folder was hard-coded, but
never created, so no cache would be used. This was to not use impure
libraries on distros other than NixOS. This commit changes that to a
path that other distros also don't have, but allows NixOS to create a
cache for its purposes, such as being compatible with Steam's
pressure-vessel.