steam: correct vulkan ICDs inside container
The ICDs will come out of their expressions with `/nix/store`-prefixed
paths, which will trip pressure-vessel up. This replaces those paths
with ones valid inside the container according to Steam's resolution
algorithm, in `/lib{64,32}`.

Furthermore, if the nvidia ICD is present, assume it's the 64 bit one,
and create a 32 bit version with `lib32` instead of `lib64`. This is
needed for 32 bit games, and only for nvidia, as other drivers have both
ICDs already.

The one thing that's problematic with this commit is that, for some
reason, `extraLibraries` aren't being propagated to the `steam-run`
environment, so running from the command line with nvidia will not work.